# tests/fixtures/assurance/backup/BACKUP-b487bf0c20ada4ec0058d954/files/tests/fixtures/api/openapi.json
{
  "components": {
    "schemas": {
      "AnswerResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_status": {
            "title": "Answer Status",
            "type": "string"
          },
          "answer_text": {
            "title": "Answer Text",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationSummary"
            },
            "title": "Citations",
            "type": "array"
          },
          "clinical_use_prohibited": {
            "default": true,
            "title": "Clinical Use Prohibited",
            "type": "boolean"
          },
          "disclaimer": {
            "default": "Synthetic portfolio demonstration only. Not for patient care. Not medical advice. Not clinically validated.",
            "title": "Disclaimer",
            "type": "string"
          },
          "generator_provider": {
            "title": "Generator Provider",
            "type": "string"
          },
          "generator_version": {
            "title": "Generator Version",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "query_id": {
            "title": "Query Id",
            "type": "string"
          },
          "rag_run_id": {
            "title": "Rag Run Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retrieval_confidence": {
            "title": "Retrieval Confidence",
            "type": "number"
          },
          "retrieval_status": {
            "title": "Retrieval Status",
            "type": "string"
          },
          "synthetic_data_only": {
            "default": true,
            "title": "Synthetic Data Only",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "query_id",
          "answer_id",
          "answer_status",
          "answer_text",
          "citations",
          "retrieval_status",
          "retrieval_confidence",
          "trace_id",
          "rag_run_id",
          "prompt_id",
          "prompt_version",
          "generator_provider",
          "generator_version"
        ],
        "title": "AnswerResponse",
        "type": "object"
      },
...
      "CitationResponse": {
        "additionalProperties": false,
        "properties": {
          "bounded_text": {
            "title": "Bounded Text",
            "type": "string"
          },
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "citation_label": {
            "title": "Citation Label",
            "type": "string"
          },
          "citation_status": {
            "title": "Citation Status",
            "type": "string"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "quoted_span": {
            "title": "Quoted Span",
            "type": "string"
          },
          "section_id": {
            "default": "",
            "title": "Section Id",
            "type": "string"
          },
          "sentence_id": {
            "default": "",
            "title": "Sentence Id",
            "type": "string"
          },
          "source_text_checksum": {
            "title": "Source Text Checksum",
            "type": "string"
          }
        },
        "required": [
          "citation_id",
          "citation_label",
          "evidence_id",
          "document_id",
          "citation_status",
          "quoted_span",
          "source_text_checksum",
          "bounded_text"
        ],
        "title": "CitationResponse",
        "type": "object"
      },
      "CitationSummary": {
        "additionalProperties": false,
        "properties": {
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "citation_label": {
            "title": "Citation Label",
            "type": "string"
          },
          "citation_status": {
            "title": "Citation Status",
            "type": "string"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "quoted_span": {
            "title": "Quoted Span",
            "type": "string"
          },
          "section_id": {
            "default": "",
            "title": "Section Id",
            "type": "string"
          },
          "sentence_id": {
            "default": "",
            "title": "Sentence Id",
            "type": "string"
          },
          "source_text_checksum": {
            "title": "Source Text Checksum",
            "type": "string"
          }
        },
        "required": [
          "citation_id",
          "citation_label",
          "evidence_id",
          "document_id",
          "citation_status",
          "quoted_span",
          "source_text_checksum"
        ],
        "title": "CitationSummary",
        "type": "object"
      },
      "EvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "bounded_snippet": {
            "title": "Bounded Snippet",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "retrieval_confidence": {
            "title": "Retrieval Confidence",
            "type": "number"
          },
          "retrieval_score": {
            "title": "Retrieval Score",
            "type": "number"
          },
          "retrieval_unit_id": {
            "title": "Retrieval Unit Id",
            "type": "string"
          },
          "section_id": {
            "default": "",
            "title": "Section Id",
            "type": "string"
          },
          "sentence_id": {
            "default": "",
            "title": "Sentence Id",
            "type": "string"
          },
          "source_extraction_run_id": {
            "title": "Source Extraction Run Id",
            "type": "string"
          },
          "source_index_id": {
            "title": "Source Index Id",
            "type": "string"
          },
          "source_preprocessing_run_id": {
            "title": "Source Preprocessing Run Id",
            "type": "string"
          },
          "text_checksum": {
            "title": "Text Checksum",
            "type": "string"
          },
          "unit_type": {
            "title": "Unit Type",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "retrieval_unit_id",
          "document_id",
          "unit_type",
          "rank",
          "retrieval_score",
          "retrieval_confidence",
          "bounded_snippet",
          "text_checksum",
          "source_preprocessing_run_id",
          "source_extraction_run_id",
          "source_index_id"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryRequest": {
        "additionalProperties": false,
        "properties": {
          "include_trace": {
            "default": false,
            "title": "Include Trace",
            "type": "boolean"
          },
          "metadata_filters": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata Filters",
            "type": "object"
          },
          "portfolio_demo_mode": {
            "default": true,
            "title": "Portfolio Demo Mode",
            "type": "boolean"
          },
          "query_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id"
          },
          "query_text": {
            "minLength": 1,
            "title": "Query Text",
            "type": "string"
          }
        },
        "required": [
          "query_text"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_status": {
            "title": "Answer Status",
            "type": "string"
          },
          "answer_text": {
            "title": "Answer Text",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationSummary"
            },
            "title": "Citations",
            "type": "array"
          },
          "clinical_use_prohibited": {
            "default": true,
            "title": "Clinical Use Prohibited",
            "type": "boolean"
          },
          "disclaimer": {
            "default": "Synthetic portfolio demonstration only. Not for patient care. Not medical advice. Not clinically validated.",
            "title": "Disclaimer",
            "type": "string"
          },
          "query_id": {
            "title": "Query Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retrieval_confidence": {
            "title": "Retrieval Confidence",
            "type": "number"
          },
          "retrieval_status": {
            "title": "Retrieval Status",
            "type": "string"
          },
          "synthetic_data_only": {
            "default": true,
            "title": "Synthetic Data Only",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "query_id",
          "answer_id",
          "answer_status",
          "answer_text",
          "citations",
          "retrieval_status",
          "retrieval_confidence",
          "trace_id"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
...
      "TraceResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "answer_status": {
            "title": "Answer Status",
            "type": "string"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Citation Ids",
            "type": "array"
          },
          "citation_validation_status": {
            "title": "Citation Validation Status",
            "type": "string"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "clinical_use_prohibited": {
            "default": true,
            "title": "Clinical Use Prohibited",
            "type": "boolean"
          },
          "excluded_retrieval_unit_ids": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Retrieval Unit Ids",
            "type": "array"
          },
          "generator_provider": {
            "title": "Generator Provider",
            "type": "string"
          },
          "generator_version": {
            "title": "Generator Version",
            "type": "string"
          },
          "groundedness_status": {
            "title": "Groundedness Status",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "query_category": {
            "title": "Query Category",
            "type": "string"
          },
          "query_id": {
            "title": "Query Id",
            "type": "string"
          },
          "retrieval_approval_id": {
            "title": "Retrieval Approval Id",
            "type": "string"
          },
          "retrieval_confidence": {
            "title": "Retrieval Confidence",
            "type": "number"
          },
          "retrieval_configuration_id": {
            "title": "Retrieval Configuration Id",
            "type": "string"
          },
          "retrieval_status": {
            "title": "Retrieval Status",
            "type": "string"
          },
          "safety_status": {
            "title": "Safety Status",
            "type": "string"
          },
          "selected_evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Selected Evidence Ids",
            "type": "array"
          },
          "synthetic_data_only": {
            "default": true,
            "title": "Synthetic Data Only",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "answer_id",
          "query_id",
          "query_category",
          "answer_status",
          "retrieval_status",
          "retrieval_configuration_id",
          "retrieval_approval_id",
          "retrieval_confidence",
          "selected_evidence_ids",
          "excluded_retrieval_unit_ids",
          "prompt_id",
          "prompt_version",
          "generator_provider",
          "generator_version",
          "claim_ids",
          "citation_ids",
          "citation_validation_status",
          "groundedness_status",
          "safety_status"
        ],
        "title": "TraceResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Synthetic-only local portfolio demonstration API. Not for patient care, not medical advice, not clinically validated, and not connected to real patient systems.",
    "title": "Healthcare Language AI Local Demonstration API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/answers/{answer_id}": {
      "get": {
        "operationId": "answer_api_v1_answers__answer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer",
        "tags": [
          "answers"
        ]
      }
    },
...
    "/api/v1/citations/{citation_id}": {
      "get": {
        "operationId": "citation_api_v1_citations__citation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "citation_id",
            "required": true,
            "schema": {
              "title": "Citation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Citation",
        "tags": [
          "citations"
        ]
      }
    },
    "/api/v1/evidence/{evidence_id}": {
      "get": {
        "operationId": "evidence_api_v1_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
...
    "/api/v1/query": {
      "post": {
        "operationId": "query_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query",
        "tags": [
          "query"
        ]
      }
    },
...
    "/api/v1/traces/{answer_id}": {
      "get": {
        "operationId": "trace_api_v1_traces__answer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trace",
        "tags": [
          "traces"
        ]
      }
    },
...
  }
}